[test] Add Test Framework
These files come from the [Unity Test Project](http://www.throwtheswitch.org/unity).
This portion of the project is licensed under the MIT license. I will
add the relevant license file to this directory, and update the readme
to reflect that. I'm choosing to include only the files relevant for the
tests so that I can more easily create a makefile for the project.