[feat] Add matrices
Added matrices as described in chapter 3. Matrices are dynamically
allocated, which resulted in a much more complex structure than the
tuples. This does allow the matrices to be arbitrarily large, which may
prove useful in later applications.

I've greatly reduced the number of commits in this chapter, as I had
felt the last two had an unnecessary amount of them. This is the
opposite extreme, with only a handful of commits.