[feat] Add Sphere skeleton
The next test introduces two new structures and a new function. It may
be necessary to immediately split off the intersection structure and
intersect function into a new set of files, and introduce a 'parent'
Shape struct to get the generality it seems the book is going to want
from the intersect function.