[feat] Add intersection functionality
Introduced new Intersection and Intersections structure, with the latter
just being a dynamically allocated collection of the former. One
important departure from the book is that the hit function now takes a
pointer to an Intersection, and returns a boolean. The intersection
pointer acts as the return value for the hit intersection, and boolean
indicates whether or not a hit has occured.