[build] Rework test makefile
I added a new 'main' directory to include the programs with their
own main function -- excluding AllTests.c, which remains in the tests
directory. This allowed for the test makefile to be reworked; it now no
longer needs object files to be explicitly listed, but retreives them
automaticially from the test and src directories.