[refactor] Add const modifiers
Added const modifiers to existing code. Following some reasonable
advice, I've added the modifier only where it both makes sense and
contributes to readability. This means that rvalues and parameters
passed by value did not get a const modifier, but functions taking
pointers and structs with pointers have const added to them if the
function does not modify that pointer.