[feat] Add sphere normals
The Sphere_NormalAt function returns the normal at a point on the
sphere. It works with transformed spheres as well. Interestingly, the
book does not give any indication that the function should check that
the passed point is actually on the sphere, so the current
implementation has no checks for that.