[refactor] Simplify Material_Lighting fuction
Broke up the Material_Lighting function using four helper functions.
This reduced the complexity of the function, and closer maintains the
single responsibility principle. Unfortunately, there is some slight
variable duplication in the helper functions, but I'm not sure of the
best way to eliminate it. Readability and maintainability is more
important than slight performance gains here.