[docs] Update readme for chapter 6
Checked off chapter 6 on the roadmap and made minor changes to other
sections. I also added another todo item, as the wrapping up portion of
chapter 6 implied that Intersection should have a reference to the
intersected object. If this turns out to not be the case, then it be
removed easily later.