[feat] Add main program utilizing camera
This program is the one described at the end of chapter 7. It functions,
but does quite a bit of intrusive memory management. This tells me that
we're going to need some better consistency for memory allocation and
deallocation. Additionally, the whole project needs some more
refactoring, as using its various functions and types is rather obtuse.
The fact that I have spend more time trying to remedy a segfault than I
did writing the basic program code tells me that the problem is fairly
serious.

This version of the code produces the image with the resolution
specified in the pseudo-code.