[build] Add warnings to main makefile
Added the same warnings to the main makefile. This generated
significantly more errors. I suspect the building in test wasn't clean
when testing. Every error and warning was simple to fix, with one
exception. The forward declaration of the Sphere type in Intersection.h
generated warnings from -pedantic. To get rid of those warnings, I
removed the typedef in Sphere.h. While this simple solution seems to
work just fine, it has leaves a counter-intuitive structure behind. 'Why
is the Sphere type defined in Intersection.h and not Sphere.h?' I'll
tolerate it for now, as the Shape class is likely going to replace the
Sphere type anyway.